feat: Add theme switching support for MultiQC and Figure components
Implement theme callbacks using Dash Patch mechanism to update figure
templates when users switch between light and dark modes. This works
for both UI mode and code mode figures without re-executing code.

Changes:
- Add dedicated theme.py modules for MultiQC and Figure components
- Register theme callbacks in component initialization
- Remove duplicate theme callback from frontend.py
- Simplify theme detection logic (remove unnecessary isinstance checks)

The callbacks use Patch to efficiently update only the template property,
switching between mantine_dark and mantine_light templates based on the
theme-store data.
