+++ depictio/dash/modules/figure_component/callbacks/__init__.py
@@ -29,9 +29,11 @@ def register_callbacks_figure_component(app):
     """
     from .core import register_core_callbacks
     from .selection import register_scatter_selection_callback
+    from .theme import register_theme_callbacks
 
     register_core_callbacks(app)
     register_scatter_selection_callback(app)
+    register_theme_callbacks(app)
 
 
 def load_design_callbacks(app):
